+++ src/package-json.ts
@@ -1,1 +1,1 @@
-export default { name: "@state-flow/core", version: "1.0.1", commit: "6b8863c9d96e7ff65eaf983b247f122628b986d5", branch: "main" };
+export default { name: "@state-flow/core", version: "1.2.0", commit: "1803cca90200b5dbd126a8409cdb7f8c9fe290e0", branch: "main" };